Feat:add function rendeSlider in CountryCard component, change data, change styles in Widgets,CountryPage components
