+++ package.json
@@ -42,6 +42,7 @@
   "devDependencies": {
     "@types/react": "^17.0.2",
     "@types/react-dom": "^17.0.1",
+    "axios": "^0.21.1",
     "typescript": "^4.2.3"
   }
 }
