+++ src/components/CountryCard/CountryCard.scss
@@ -3,6 +3,7 @@
     margin: 0 auto;
     border: 0 solid #e2e7ea;
     box-sizing: border-box;
+    flex-grow: 7;
 }
 
 .country-wrapper--description {
