+++ src/components/CurrencyWidget/CurrencyWidget.scss
@@ -1,5 +1,5 @@
 .currency {
-  width: 60%;
+  width: 100%;
   color: whitesmoke;
   margin: 1rem;
 }
