+++ src/components/TimeWidget/TimeWidget.scss
@@ -1,5 +1,5 @@
 .date-time {
-  width: 60%;
+  width: 100%;
   color: whitesmoke;
   margin: 1rem;
 }
