+++ src/components/WeatherWidget/WeatherWidget.scss
@@ -1,5 +1,5 @@
 .weather {
-    width: 60%;
+    width: 100%;
     color: whitesmoke;
     margin: 1rem;
