+++ src/components/WidgetsBox/WidgetsBox.scss
@@ -1,3 +1,4 @@
 .widget_box {
+   flex-grow: 3;
    width: 30%;
 }
