+++ src/data/dataSlider.ts
@@ -1,0 +1,22 @@
+const dataCarousel = [
+    {
+      imgURL:
+        "https://wikiway.com/upload/hl-photo/00c/452/kievo-pecherskaya-lavra_130.jpg",
+      description:
+        "Киево-Печерская лавра.Легендарный историко-архитектурный объект города Киев. Дата возведения – 1051 год.",
+    },
+    {
+      imgURL: "https://wikiway.com/upload/hl-photo/b06/378/bukovel_124.jpg",
+      description:
+        " Буковель – один из самых известных украинских горнолыжных курортов. Он расположен на высоте более 900 метров рядом с селом Поляница Ивано-Франковской области.",
+    },
+    {
+      imgURL:
+        "https://wikiway.com/upload/hl-photo/e92/be1/chernobylskaya-aes_75.jpg",
+      description:
+        " Буковель – один из самых известных украинских горнолыжных курортов. Он расположен на высоте более 900 метров рядом с селом Поляница Ивано-Франковской области.",
+    },
+  ];
+
+
+  export default dataCarousel
