+++ src/pages/CountryPage/CountryPage.scss
@@ -1,5 +1,6 @@
-.country_page__container {   
+.country_page {   
         display: flex;
-        justify-content: flex-end;
+        flex-direction: row;
+        flex-wrap: nowrap;
         margin: 1rem 2rem;
 }
